chore: merge lexer into tokenizer file
